refactor: consolidate alias rewriting functions into expression_handlers
- Move rewrite_aliases_in_stmt and helper functions from stdlib_normalization.rs
  to expression_handlers.rs, making them public for reuse
- Replace existing rewrite_aliases_in_expr with more powerful version that:
  * Handles both Load and Store contexts (fixes assignment target rewriting)
  * Converts dotted names (e.g., PyPath → pathlib.Path) for Load contexts
  * Properly rewrites names in all expression types including f-strings
- Remove duplicate implementations from bundler.rs
- Fix critical bug where global statements in functions weren't being rewritten
  correctly by processing entire function definition instead of individual statements

This eliminates code duplication and provides a single, robust implementation
for alias rewriting that correctly handles all cases including global statements
and assignment targets.
