+++ crates/cribo/src/stdlib_normalization.rs
@@ -1,12 +1,10 @@
 use log::debug;
-use ruff_python_ast::{
-    ExceptHandler, Expr, ExprContext, Identifier, ModModule, Stmt, StmtClassDef, StmtFunctionDef,
-    StmtImport,
-};
+use ruff_python_ast::{ExprContext, Identifier, ModModule, Stmt, StmtImport};
 use ruff_text_size::TextRange;
 
 use crate::{
     ast_builder::{expressions, statements},
+    code_generator::expression_handlers,
     side_effects::is_safe_stdlib_module,
     types::{FxIndexMap, FxIndexSet},
 };
@@ -111,7 +109,7 @@ impl StdlibNormalizer {
                         _ => format!("{:?}", std::mem::discriminant(stmt)),
                     };
                     debug!("Rewriting aliases in statement at index {idx}: {stmt_type}");
-                    self.rewrite_aliases_in_stmt(stmt, &alias_to_canonical);
+                    expression_handlers::rewrite_aliases_in_stmt(stmt, &alias_to_canonical);
                 }
             }
         }
@@ -313,502 +311,7 @@ impl StdlibNormalizer {
         }
     }
 
-    /// Recursively rewrite aliases in statements
-    fn rewrite_aliases_in_stmt(
-        &self,
-        stmt: &mut Stmt,
-        alias_to_canonical: &FxIndexMap<String, String>,
-    ) {
-        match stmt {
-            Stmt::Expr(expr_stmt) => {
-                Self::rewrite_aliases_in_expr(&mut expr_stmt.value, alias_to_canonical);
-            }
-            Stmt::Assign(assign) => {
-                Self::rewrite_aliases_in_expr(&mut assign.value, alias_to_canonical);
-                for target in &mut assign.targets {
-                    Self::rewrite_aliases_in_expr(target, alias_to_canonical);
-                }
-            }
-            Stmt::Return(return_stmt) => {
-                debug!("Rewriting aliases in return statement");
-                if let Some(ref mut value) = return_stmt.value {
-                    Self::rewrite_aliases_in_expr(value, alias_to_canonical);
-                }
-            }
-            Stmt::FunctionDef(func_def) => {
-                self.rewrite_aliases_in_function(func_def, alias_to_canonical);
-            }
-            Stmt::ClassDef(class_def) => {
-                self.rewrite_aliases_in_class(class_def, alias_to_canonical);
-            }
-            Stmt::AnnAssign(ann_assign) => {
-                // Rewrite the annotation
-                Self::rewrite_aliases_in_expr(&mut ann_assign.annotation, alias_to_canonical);
-                // Rewrite the target
-                Self::rewrite_aliases_in_expr(&mut ann_assign.target, alias_to_canonical);
-                // Rewrite the value if present
-                if let Some(ref mut value) = ann_assign.value {
-                    Self::rewrite_aliases_in_expr(value, alias_to_canonical);
-                }
-            }
-            Stmt::If(if_stmt) => {
-                Self::rewrite_aliases_in_expr(&mut if_stmt.test, alias_to_canonical);
-                for stmt in &mut if_stmt.body {
-                    self.rewrite_aliases_in_stmt(stmt, alias_to_canonical);
-                }
-                for clause in &mut if_stmt.elif_else_clauses {
-                    if let Some(ref mut condition) = clause.test {
-                        Self::rewrite_aliases_in_expr(condition, alias_to_canonical);
-                    }
-                    for stmt in &mut clause.body {
-                        self.rewrite_aliases_in_stmt(stmt, alias_to_canonical);
-                    }
-                }
-            }
-            Stmt::While(while_stmt) => {
-                Self::rewrite_aliases_in_expr(&mut while_stmt.test, alias_to_canonical);
-                for stmt in &mut while_stmt.body {
-                    self.rewrite_aliases_in_stmt(stmt, alias_to_canonical);
-                }
-                for stmt in &mut while_stmt.orelse {
-                    self.rewrite_aliases_in_stmt(stmt, alias_to_canonical);
-                }
-            }
-            Stmt::For(for_stmt) => {
-                Self::rewrite_aliases_in_expr(&mut for_stmt.iter, alias_to_canonical);
-                for stmt in &mut for_stmt.body {
-                    self.rewrite_aliases_in_stmt(stmt, alias_to_canonical);
-                }
-                for stmt in &mut for_stmt.orelse {
-                    self.rewrite_aliases_in_stmt(stmt, alias_to_canonical);
-                }
-            }
-            Stmt::With(with_stmt) => {
-                for item in &mut with_stmt.items {
-                    Self::rewrite_aliases_in_expr(&mut item.context_expr, alias_to_canonical);
-                }
-                for stmt in &mut with_stmt.body {
-                    self.rewrite_aliases_in_stmt(stmt, alias_to_canonical);
-                }
-            }
-            Stmt::Try(try_stmt) => {
-                for stmt in &mut try_stmt.body {
-                    self.rewrite_aliases_in_stmt(stmt, alias_to_canonical);
-                }
-                for handler in &mut try_stmt.handlers {
-                    self.rewrite_aliases_in_except_handler(handler, alias_to_canonical);
-                }
-                for stmt in &mut try_stmt.orelse {
-                    self.rewrite_aliases_in_stmt(stmt, alias_to_canonical);
-                }
-                for stmt in &mut try_stmt.finalbody {
-                    self.rewrite_aliases_in_stmt(stmt, alias_to_canonical);
-                }
-            }
-            Stmt::AugAssign(aug_assign) => {
-                Self::rewrite_aliases_in_expr(&mut aug_assign.target, alias_to_canonical);
-                Self::rewrite_aliases_in_expr(&mut aug_assign.value, alias_to_canonical);
-            }
-            Stmt::Raise(raise_stmt) => {
-                if let Some(ref mut exc) = raise_stmt.exc {
-                    Self::rewrite_aliases_in_expr(exc, alias_to_canonical);
-                }
-                if let Some(ref mut cause) = raise_stmt.cause {
-                    Self::rewrite_aliases_in_expr(cause, alias_to_canonical);
-                }
-            }
-            Stmt::Assert(assert_stmt) => {
-                Self::rewrite_aliases_in_expr(&mut assert_stmt.test, alias_to_canonical);
-                if let Some(ref mut msg) = assert_stmt.msg {
-                    Self::rewrite_aliases_in_expr(msg, alias_to_canonical);
-                }
-            }
-            Stmt::Delete(delete_stmt) => {
-                for target in &mut delete_stmt.targets {
-                    Self::rewrite_aliases_in_expr(target, alias_to_canonical);
-                }
-            }
-            Stmt::Global(_)
-            | Stmt::Nonlocal(_)
-            | Stmt::Pass(_)
-            | Stmt::Break(_)
-            | Stmt::Continue(_) => {
-                // These statements don't contain expressions to rewrite
-            }
-            // Handle other statement types as needed
-            _ => {
-                debug!(
-                    "Unhandled statement type in rewrite_aliases_in_stmt: {:?}",
-                    std::mem::discriminant(stmt)
-                );
-            }
-        }
-    }
-
-    /// Rewrite aliases in expressions
-    fn rewrite_aliases_in_expr(expr: &mut Expr, alias_to_canonical: &FxIndexMap<String, String>) {
-        match expr {
-            Expr::Name(name_expr) if matches!(name_expr.ctx, ExprContext::Load) => {
-                // Check if this is an aliased import that should be rewritten
-                if let Some(canonical) = alias_to_canonical.get(name_expr.id.as_str()) {
-                    debug!(
-                        "Rewriting name '{}' to '{}'",
-                        name_expr.id.as_str(),
-                        canonical
-                    );
-                    // Replace simple name with the canonical form
-                    if canonical.contains('.') {
-                        // Convert to attribute access (e.g., pathlib.Path)
-                        let parts: Vec<&str> = canonical.split('.').collect();
-                        *expr = expressions::dotted_name(&parts, ExprContext::Load);
-                    } else {
-                        // Simple module name
-                        name_expr.id = canonical.clone().into();
-                    }
-                }
-            }
-            Expr::Attribute(attr_expr) => {
-                // Check if the base is an aliased module
-                if let Expr::Name(name_expr) = &mut *attr_expr.value {
-                    if let Some(canonical) = alias_to_canonical.get(name_expr.id.as_str()) {
-                        debug!(
-                            "Rewriting attribute base '{}' to '{}' in attribute expression",
-                            name_expr.id.as_str(),
-                            canonical
-                        );
-                        name_expr.id = canonical.clone().into();
-                    }
-                } else {
-                    debug!(
-                        "Attribute base is not a Name expression, it's {:?}",
-                        std::mem::discriminant(&*attr_expr.value)
-                    );
-                }
-                // Recursively process the value
-                Self::rewrite_aliases_in_expr(&mut attr_expr.value, alias_to_canonical);
-            }
-            Expr::Call(call_expr) => {
-                // Debug the function being called
-                if let Expr::Name(name) = &*call_expr.func {
-                    debug!("Call expression with function name: {}", name.id.as_str());
-                } else if let Expr::Attribute(attr) = &*call_expr.func
-                    && let Expr::Name(base) = &*attr.value
-                {
-                    debug!(
-                        "Call expression with attribute: {}.{}",
-                        base.id.as_str(),
-                        attr.attr.as_str()
-                    );
-                }
-
-                Self::rewrite_aliases_in_expr(&mut call_expr.func, alias_to_canonical);
-                for (i, arg) in call_expr.arguments.args.iter_mut().enumerate() {
-                    debug!("  Rewriting call arg {i}");
-                    Self::rewrite_aliases_in_expr(arg, alias_to_canonical);
-                }
-                for keyword in &mut call_expr.arguments.keywords {
-                    Self::rewrite_aliases_in_expr(&mut keyword.value, alias_to_canonical);
-                }
-            }
-            // Handle other expression types recursively
-            Expr::List(list_expr) => {
-                debug!("Rewriting aliases in list expression");
-                for elem in &mut list_expr.elts {
-                    Self::rewrite_aliases_in_expr(elem, alias_to_canonical);
-                }
-            }
-            Expr::Tuple(tuple_expr) => {
-                debug!("Rewriting aliases in tuple expression");
-                for elem in &mut tuple_expr.elts {
-                    Self::rewrite_aliases_in_expr(elem, alias_to_canonical);
-                }
-            }
-            Expr::Subscript(subscript_expr) => {
-                debug!("Rewriting aliases in subscript expression");
-                Self::rewrite_aliases_in_expr(&mut subscript_expr.value, alias_to_canonical);
-                Self::rewrite_aliases_in_expr(&mut subscript_expr.slice, alias_to_canonical);
-            }
-            Expr::BinOp(binop) => {
-                Self::rewrite_aliases_in_expr(&mut binop.left, alias_to_canonical);
-                Self::rewrite_aliases_in_expr(&mut binop.right, alias_to_canonical);
-            }
-            Expr::UnaryOp(unaryop) => {
-                Self::rewrite_aliases_in_expr(&mut unaryop.operand, alias_to_canonical);
-            }
-            Expr::Compare(compare) => {
-                Self::rewrite_aliases_in_expr(&mut compare.left, alias_to_canonical);
-                for comparator in &mut compare.comparators {
-                    Self::rewrite_aliases_in_expr(comparator, alias_to_canonical);
-                }
-            }
-            Expr::BoolOp(boolop) => {
-                for value in &mut boolop.values {
-                    Self::rewrite_aliases_in_expr(value, alias_to_canonical);
-                }
-            }
-            Expr::Dict(dict) => {
-                for item in &mut dict.items {
-                    if let Some(ref mut key) = item.key {
-                        Self::rewrite_aliases_in_expr(key, alias_to_canonical);
-                    }
-                    Self::rewrite_aliases_in_expr(&mut item.value, alias_to_canonical);
-                }
-            }
-            Expr::Set(set) => {
-                for elem in &mut set.elts {
-                    Self::rewrite_aliases_in_expr(elem, alias_to_canonical);
-                }
-            }
-            Expr::ListComp(comp) => {
-                Self::rewrite_aliases_in_expr(&mut comp.elt, alias_to_canonical);
-                for generator in &mut comp.generators {
-                    Self::rewrite_aliases_in_expr(&mut generator.iter, alias_to_canonical);
-                    for if_clause in &mut generator.ifs {
-                        Self::rewrite_aliases_in_expr(if_clause, alias_to_canonical);
-                    }
-                }
-            }
-            Expr::SetComp(comp) => {
-                Self::rewrite_aliases_in_expr(&mut comp.elt, alias_to_canonical);
-                for generator in &mut comp.generators {
-                    Self::rewrite_aliases_in_expr(&mut generator.iter, alias_to_canonical);
-                    for if_clause in &mut generator.ifs {
-                        Self::rewrite_aliases_in_expr(if_clause, alias_to_canonical);
-                    }
-                }
-            }
-            Expr::DictComp(comp) => {
-                Self::rewrite_aliases_in_expr(&mut comp.key, alias_to_canonical);
-                Self::rewrite_aliases_in_expr(&mut comp.value, alias_to_canonical);
-                for generator in &mut comp.generators {
-                    Self::rewrite_aliases_in_expr(&mut generator.iter, alias_to_canonical);
-                    for if_clause in &mut generator.ifs {
-                        Self::rewrite_aliases_in_expr(if_clause, alias_to_canonical);
-                    }
-                }
-            }
-            Expr::Generator(comp) => {
-                Self::rewrite_aliases_in_expr(&mut comp.elt, alias_to_canonical);
-                for generator in &mut comp.generators {
-                    Self::rewrite_aliases_in_expr(&mut generator.iter, alias_to_canonical);
-                    for if_clause in &mut generator.ifs {
-                        Self::rewrite_aliases_in_expr(if_clause, alias_to_canonical);
-                    }
-                }
-            }
-            Expr::Lambda(lambda) => {
-                // Lambda parameters might have annotations
-                if let Some(ref mut params) = lambda.parameters {
-                    for param in &mut params.posonlyargs {
-                        if let Some(ref mut annotation) = param.parameter.annotation {
-                            Self::rewrite_aliases_in_expr(annotation, alias_to_canonical);
-                        }
-                    }
-                    for param in &mut params.args {
-                        if let Some(ref mut annotation) = param.parameter.annotation {
-                            Self::rewrite_aliases_in_expr(annotation, alias_to_canonical);
-                        }
-                    }
-                    for param in &mut params.kwonlyargs {
-                        if let Some(ref mut annotation) = param.parameter.annotation {
-                            Self::rewrite_aliases_in_expr(annotation, alias_to_canonical);
-                        }
-                    }
-                }
-                // Process the body
-                Self::rewrite_aliases_in_expr(&mut lambda.body, alias_to_canonical);
-            }
-            Expr::If(ifexp) => {
-                Self::rewrite_aliases_in_expr(&mut ifexp.test, alias_to_canonical);
-                Self::rewrite_aliases_in_expr(&mut ifexp.body, alias_to_canonical);
-                Self::rewrite_aliases_in_expr(&mut ifexp.orelse, alias_to_canonical);
-            }
-            Expr::Yield(yield_expr) => {
-                if let Some(ref mut value) = yield_expr.value {
-                    Self::rewrite_aliases_in_expr(value, alias_to_canonical);
-                }
-            }
-            Expr::YieldFrom(yield_from) => {
-                Self::rewrite_aliases_in_expr(&mut yield_from.value, alias_to_canonical);
-            }
-            Expr::Await(await_expr) => {
-                Self::rewrite_aliases_in_expr(&mut await_expr.value, alias_to_canonical);
-            }
-            Expr::Starred(starred) => {
-                Self::rewrite_aliases_in_expr(&mut starred.value, alias_to_canonical);
-            }
-            Expr::Slice(slice) => {
-                if let Some(ref mut lower) = slice.lower {
-                    Self::rewrite_aliases_in_expr(lower, alias_to_canonical);
-                }
-                if let Some(ref mut upper) = slice.upper {
-                    Self::rewrite_aliases_in_expr(upper, alias_to_canonical);
-                }
-                if let Some(ref mut step) = slice.step {
-                    Self::rewrite_aliases_in_expr(step, alias_to_canonical);
-                }
-            }
-            Expr::Named(named) => {
-                Self::rewrite_aliases_in_expr(&mut named.value, alias_to_canonical);
-            }
-            // Literals don't need rewriting
-            Expr::StringLiteral(_)
-            | Expr::BytesLiteral(_)
-            | Expr::NumberLiteral(_)
-            | Expr::BooleanLiteral(_)
-            | Expr::NoneLiteral(_)
-            | Expr::EllipsisLiteral(_) => {
-                // No aliases to rewrite in literals
-            }
-            _ => {
-                debug!(
-                    "Unhandled expression type in rewrite_aliases_in_expr: {:?}",
-                    std::mem::discriminant(expr)
-                );
-            }
-        }
-    }
-
-    /// Rewrite aliases in exception handlers
-    fn rewrite_aliases_in_except_handler(
-        &self,
-        handler: &mut ExceptHandler,
-        alias_to_canonical: &FxIndexMap<String, String>,
-    ) {
-        match handler {
-            ExceptHandler::ExceptHandler(except_handler) => {
-                if let Some(ref mut type_) = except_handler.type_ {
-                    Self::rewrite_aliases_in_expr(type_, alias_to_canonical);
-                }
-                for stmt in &mut except_handler.body {
-                    self.rewrite_aliases_in_stmt(stmt, alias_to_canonical);
-                }
-            }
-        }
-    }
-
-    /// Rewrite aliases in function definitions
-    fn rewrite_aliases_in_function(
-        &self,
-        func_def: &mut StmtFunctionDef,
-        alias_to_canonical: &FxIndexMap<String, String>,
-    ) {
-        debug!("Rewriting aliases in function: {}", func_def.name.as_str());
-
-        // Rewrite parameter annotations
-        for param in &mut func_def.parameters.posonlyargs {
-            if let Some(ref mut annotation) = param.parameter.annotation {
-                Self::rewrite_aliases_in_expr(annotation, alias_to_canonical);
-            }
-        }
-        for param in &mut func_def.parameters.args {
-            if let Some(ref mut annotation) = param.parameter.annotation {
-                Self::rewrite_aliases_in_expr(annotation, alias_to_canonical);
-            }
-        }
-        for param in &mut func_def.parameters.kwonlyargs {
-            if let Some(ref mut annotation) = param.parameter.annotation {
-                Self::rewrite_aliases_in_expr(annotation, alias_to_canonical);
-            }
-        }
-        if let Some(ref mut vararg) = func_def.parameters.vararg
-            && let Some(ref mut annotation) = vararg.annotation
-        {
-            Self::rewrite_aliases_in_expr(annotation, alias_to_canonical);
-        }
-        if let Some(ref mut kwarg) = func_def.parameters.kwarg
-            && let Some(ref mut annotation) = kwarg.annotation
-        {
-            Self::rewrite_aliases_in_expr(annotation, alias_to_canonical);
-        }
-
-        // Rewrite return type annotation
-        if let Some(ref mut returns) = func_def.returns {
-            Self::rewrite_aliases_in_expr(returns, alias_to_canonical);
-        }
-
-        // First handle global statements specially
-        self.rewrite_global_statements_in_function(func_def, alias_to_canonical);
-
-        // Then rewrite the rest of the function body
-        for (idx, stmt) in func_def.body.iter_mut().enumerate() {
-            match stmt {
-                Stmt::Global(_) => {
-                    // Already handled above
-                }
-                _ => {
-                    debug!(
-                        "  Rewriting aliases in function body statement {}: {:?}",
-                        idx,
-                        std::mem::discriminant(stmt)
-                    );
-                    self.rewrite_aliases_in_stmt(stmt, alias_to_canonical);
-                }
-            }
-        }
-    }
-
-    /// Rewrite only global statements in function, not other references
-    fn rewrite_global_statements_in_function(
-        &self,
-        func_def: &mut StmtFunctionDef,
-        alias_to_canonical: &FxIndexMap<String, String>,
-    ) {
-        for stmt in &mut func_def.body {
-            Self::rewrite_global_statements_only(stmt, alias_to_canonical);
-        }
-    }
-
-    /// Recursively rewrite only global statements, not other name references
-    fn rewrite_global_statements_only(
-        stmt: &mut Stmt,
-        alias_to_canonical: &FxIndexMap<String, String>,
-    ) {
-        match stmt {
-            Stmt::Global(global_stmt) => {
-                // Apply renames to global variable names
-                for name in &mut global_stmt.names {
-                    let name_str = name.as_str();
-                    if let Some(new_name) = alias_to_canonical.get(name_str) {
-                        debug!("Rewriting global statement variable '{name_str}' to '{new_name}'");
-                        *name = Identifier::new(new_name, TextRange::default());
-                    }
-                }
-            }
-            // For control flow statements, recurse into their bodies
-            Stmt::If(if_stmt) => {
-                for stmt in &mut if_stmt.body {
-                    Self::rewrite_global_statements_only(stmt, alias_to_canonical);
-                }
-                for clause in &mut if_stmt.elif_else_clauses {
-                    for stmt in &mut clause.body {
-                        Self::rewrite_global_statements_only(stmt, alias_to_canonical);
-                    }
-                }
-            }
-            // Handle other control flow statements similarly...
-            _ => {}
-        }
-    }
-
-    /// Rewrite aliases in class definitions
-    fn rewrite_aliases_in_class(
-        &self,
-        class_def: &mut StmtClassDef,
-        alias_to_canonical: &FxIndexMap<String, String>,
-    ) {
-        // Rewrite base classes
-        if let Some(arguments) = &mut class_def.arguments {
-            for base in &mut arguments.args {
-                Self::rewrite_aliases_in_expr(base, alias_to_canonical);
-            }
-        }
-
-        // Rewrite class body
-        for stmt in &mut class_def.body {
-            self.rewrite_aliases_in_stmt(stmt, alias_to_canonical);
-        }
-    }
+    // All the rewrite functions have been moved to expression_handlers.rs
 
     /// Process names from an import-from statement and collect imports and exports
     fn process_import_from_names(
